generational verification working, so it works!
the final piece is in place. generational verification ensures overall
correctness and somewhat manages combinatoric possibility explosion by
culling clearly unfit offspring during the process

however, there is no mechanism for recognizing equivalent states, so
many of the solutions/offspring in a given generation are
redundant. that is a harder problem...